+++ Tests/PuppyTests/OSLoggerTests.swift
@@ -1,7 +1,6 @@
 #if canImport(Darwin)
-
 import XCTest
-@testable import Puppy
+import Puppy
 
 final class OSLoggerTests: XCTestCase {